api/pyproject.toml
[project]
name = "wiki-rag-api"
version = "0.2.0"
description = "API service for wiki-rag on Azure"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "psycopg[binary]>=3.1",
  "pgvector>=0.2",
  "llama-index-core>=0.10",
  "llama-index-embeddings-openai>=0.1",
  "llama-index-vector-stores-postgres>=0.1",
  "openai>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
